renamed to style files to style. Better name
