+++ src/App.css
@@ -9,7 +9,7 @@
 
 /* Container Home */
 .container-home {
-  margin:2rem auto;
+  margin:4rem auto;
   height: 20%;
   background-color: rgb(44, 43, 43);
   display: flex;
@@ -19,6 +19,7 @@
   padding: 2rem 0rem;
   align-items: center;
   width: 60%;
+
 }
 
 /* Typography */
@@ -63,7 +64,8 @@ h5 {
 }
 .home-link{
   font-size: large;
-  background-color: -webkit-linear-gradient(left, #4e44ce, #35aee2) !important;
+  margin:0 auto;
+  background: -webkit-linear-gradient(left, #a944ce, #35aee2) !important;
 }
 .home-link a{
   text-decoration: none;
