+++ src/components/FetchNft/styles.ts
@@ -1,0 +1,3 @@
+export const linkStyle = {
+  textDecoration: "none",
+};
